pyproject.toml
[project]
name = "equation-maker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.9"
dependencies = [
    "fastapi[standard]>=0.121.0",
    "sqlmodel>=0.0.27",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pyjwt>=2.10.1",
]

[tool.ruff]
line-length = 120
